+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Module [AHAB](./docs/ahab/Ahab.md)
     - Added gate length as a parameter for operator `>` in note mode (#420)
+    - Added reset input for tick counter (#429)
     - Fixed crash on operator `<` when using whithout maximum value set (#425)
 - Module [MB](./docs/trial-and-error/Mb.md)
     - Added fuzzy search to v1 browser (similar to the default module browser)